pyproject.toml
[tool.poetry]
name = "lebedev"
version = "0.1.0"
description = ""
authors = ["Telekurysh <mic.mamaev@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
